+++ app.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const app = express();
-const { router } = require('./routers/main');
+const mainRouter = require('./routers/main');
+const productsRouter = require('./routers/products');
 
 const port = 3000;
 
@@ -12,7 +13,8 @@ app.use(express.urlencoded({ extended: false }));
 
 app.use(express.json());
 
-app.use('/', router);
+app.use('/', mainRouter);
+app.use('/', productsRouter);
 
 app.listen(port, () => {
     console.log(`App listening on port http://localhost:${port}`);
