+++ controllers/mainController.js
@@ -1,6 +1,3 @@
-const products = require('../dataBase/products');
-const { cart } = require("../dataBase/cart");
-
 const mainController = {
   home: (req, res) => {
     res.render('home');
@@ -14,18 +11,8 @@ const mainController = {
   billing: (req, res) => {
     res.render('billing');
   },
-  productDetail: (req, res) => {
-    const productSearch = products.find((product)=> product.id === req.params.id)
-    res.render('products/product-detail',{data: products, search:productSearch});
-  },
-  productsCart: (req, res) => {
-    res.render('products/products-cart', { cart });
-  },
   redirect: (req, res) => {
     res.redirect('/');
-  },
-  loadSandwich: (req, res) => {
-    res.render('load-sandwich');
   }
 }
